restaurant-customer-api/src/main/java/com/yosep/restaurant/interfaces/CategoryController.java
package com.yosep.restaurant.interfaces;

import java.net.URI;
import java.net.URISyntaxException;
import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import com.yosep.restaurant.application.CategoryService;
import com.yosep.restaurant.domain.Category;

@RestController
public class CategoryController {
	@Autowired
	CategoryService categoryService;
	
	@GetMapping("/categories")
	public List<Category> list() {
		List<Category> categories = categoryService.getCategories();
		
		return categories;
	}
}